requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
gitdb2==2.0.3
GitPython==2.1.7
idna==2.6
python-slugify==1.2.4
PyYAML==3.12
requests==2.18.4
smmap2==2.0.3
Unidecode==0.4.21
urllib3==1.22
terminaltables==3.1.0
inflection==0.3.1
